+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 2.5.1
+#### Changes
+- Add showBannerView / hideBannerView API
+- Fix Bugs
+#### Integration Bidmad SDK version 
+- Android : bidmad-androidx-g 1.0.4
+- iOS : BidmadSDK.framework 2.3.0
+
 # Version 2.5.0
 #### Changes
 - Change Bidmad Interface structure